+++ api/mypage/mypage.ts
