+++ api/queryClient.ts
@@ -1,4 +1,3 @@
-// api/queryClient.ts
 import { QueryClient } from '@tanstack/react-query';
 
 const queryClient = new QueryClient({
